backend/requirements.txt
fastapi
uvicorn
pandas
numpy
jinja2
weasyprint
pydantic
python-multipart
python-dotenv
pyjwt
scipy
openpyxl
annotated-types
requests
schedule
